option to specify color from the GUI

+++ BehaveAndScanJing/struct_def.cs
@@ -44,8 +44,8 @@ namespace BehaveAndScanSPIM
         public double flash_freq;
         public double flash_dur;
         public double flash_iti;
-        public double flash_col1;
-        public double flash_col2;
+        public int flash_col1;
+        public int flash_col2;
 
     }
